Version .5 Commit
///Additions & Changes///
-min_height and min_width are now experimental in that the slideshow now
conforms to the CSS settings of the .window_frame element. This gives a
user easier control of how the show is sized in the users browser window
but now may conflict with these values.
-Above values are now set to zero by default for above reasons.
-New preview element/behavior. A click on a nav item now displays an
HTML based preview text for that pane. This is passed when a user
creates the pane by attaching a 'data-title' attribue.

///Bug Fixes///
-Fixed bug where win_h and win_w would not update properly on a window
resize.
-Fixed bug where current pane nav element would be reset to 0 on a
window resize.
-Fixed a bug where pane widths and heights were not adjusting to the new
container size on a window resize.